requirements.txt
asgiref==3.3.1
Django==2.2.12
django-filter==2.4.0
django-widget-tweaks==1.4.8
pytz==2020.4
sqlparse==0.4.1
